move scheduler to its own package
